Разделяет utils на render и common
